+++ psgscoring/constants.py
@@ -93,6 +93,7 @@ def _profile_to_legacy_dict(profile) -> dict:
         "THERMISTOR_GATE":          pp.thermistor_gate,
         "HYPOPNEA_FORCE_LINEARISATION": pp.hypopnea_force_linearisation,
         "MAX_EVENTS_PER_DESATURATION": pp.max_events_per_desaturation,
+        "RIP_QUALITY_SCALE_FREE": pp.rip_quality_scale_free,
         "HYPOXIC_BURDEN_LOCAL_BASELINE": pp.hypoxic_burden_local_baseline,
         "HYPOPNEA_AROUSAL_LATENCY": pp.hypopnea_arousal_latency_grading,
         "RULE1A_AROUSAL_ENABLED": pp.rule1a_arousal_enabled,